% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhc_multinomial.R
\name{bhc.multinomial}
\alias{bhc.multinomial}
\title{Bayesian hierarchical clustering for multinomial variates}
\usage{
bhc.multinomial(x, alpha = 1, beta = rep(1, ncol(x)))
}
\arguments{
\item{x}{An integer matrix with rows representing observations
to cluster.}

\item{alpha}{Dirichlet process hyperparameter.}

\item{beta}{Dirichlet distribution hyperparameter.}
}
\description{
Bayesian hierarchical clustering for multinomial variates
}
\details{
Data observations are assumed to follow a multinomial
distribution. A Dirichlet distribution is used as the prior
distribution for the multinomial probabilities.
}
\examples{
data(snakediet)
x = xtabs(snakediet$count ~ snakediet$species + snakediet$food)
res = bhc.multinomial(x)
}
