% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhc_normal.R
\name{bhc.normal}
\alias{bhc.normal}
\title{Bayesian hierarchical clustering for normal variates}
\usage{
bhc.normal(
  x,
  alpha = 1,
  nu = ncol(x),
  kappa = 1,
  mu = colMeans(x, na.rm = TRUE),
  lambda = var(x, na.rm = TRUE) * ncol(x)
)
}
\arguments{
\item{x}{A numeric matrix with rows representing observations
to cluster.}

\item{alpha}{Dirichlet process hyperparameter.}

\item{nu}{Normal-Inverse-Wishart (NIV) prior degrees of freedom.}

\item{kappa}{NIV prior number of observations.}

\item{mu}{NIV prior mean.}

\item{lambda}{NIV prior sum of squares matrix.}
}
\description{
Bayesian hierarchical clustering for normal variates
}
\details{
Data observations are assumed to follow a multivariate normal
distribution. A normal-inverse-wishart distribution is used as the prior
distribution for the mean and covariance.
}
