% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhc_normal.R
\name{fit.bhc.normal}
\alias{fit.bhc.normal}
\title{Bayesian hierarchical clustering for normal variates}
\usage{
fit.bhc.normal(x, alpha = 1)
}
\arguments{
\item{x}{A numeric matrix with rows representing observations
to cluster.}

\item{alpha}{Dirichlet process hyperparameter.}
}
\description{
Bayesian hierarchical clustering for normal variates
}
\details{
Optimizes the prior sum of squares matrix while holding
the prior mean fixed to the observed mean. All other hyperparameters
are also held fixed. Notably, the Dirichlet process parameter
is held fixed to the value supplied in the argument.
}
